+++ apps/analog-app/vite.config.ts
@@ -32,14 +32,18 @@ export default defineConfig(({ mode }) => {
         additionalPagesDirs: ['/libs/shared/feature'],
         additionalAPIDirs: ['/libs/shared/feature/src/api'],
         prerender: {
-          routes: ['/', '/cart', '/shipping'],
+          routes: ['/', '/cart', '/shipping', '/client'],
           sitemap: {
             host: base,
           },
         },
         vite: {
           inlineStylesExtension: 'scss',
+          experimental: {
+            supportAnalogFormat: true,
+          },
         },
+        liveReload: true,
       }),
       nxViteTsPaths(),
       visualizer() as Plugin,